chore(use-case): refresh tests for latest cdk and constructs (#1237)
* Refresh use case tests

* Update README, trigger build

* Update zipped web site name
