+++ .viperlightignore
@@ -183,4 +183,4 @@ source/patterns/@aws-solutions-constructs/aws-route53-alb/test/integ.r53alb-depl
 source/patterns/@aws-solutions-constructs/aws-route53-alb/test/integ.r53alb-deployPublicApiNewAlb.js.snapshot/r53alb-deployPublicApiNewAlb.template.json:214
 source/patterns/@aws-solutions-constructs/aws-route53-alb/test/integ.r53alb-deployPublicApiNewAlb.js.snapshot/tree.json:269
 # compressed test web site
-source/use_cases/aws-s3-static-website/test/integ.basic-deployment.js.snapshot/asset.3322b7049fb0ed2b7cbb644a2ada8d1116ff80c32dca89e6ada846b5de26f961.zip
+source/use_cases/aws-s3-static-website/test/integ.basic-deployment.js.snapshot/asset.eafa2d6c6a55bda16b1603003f1721e2f66a024c0c31d6665ee1c5ff59b565d9.zip
