+++ source/use_cases/aws-s3-static-website/test/integ.basic-deployment.js.snapshot/StaticWebsiteStackIntegDefaultTestDeployAssertD6A7B854.assets.json
@@ -1,5 +1,5 @@
 {
-  "version": "36.0.0",
+  "version": "38.0.1",
   "files": {
     "21fbb51d7b23f6a6c262b46a9caee79d744a3ac019fd45422d988b96d44b2a22": {
       "source": {
