+++ source/use_cases/aws-s3-static-website/test/integ.basic-deployment.js.snapshot/cdk.out
@@ -1,1 +1,1 @@
-{"version":"36.0.0"}
+{"version":"38.0.1"}
