+++ source/use_cases/aws-s3-static-website/test/integ.basic-deployment.js.snapshot/integ.json
@@ -1,5 +1,5 @@
 {
-  "version": "36.0.0",
+  "version": "38.0.1",
   "testCases": {
     "StaticWebsiteStack/Integ/DefaultTest": {
       "stacks": [
